refactor(workflow): split task outcome service
